bug fixing

+++ src/com/sirma/SortedByPhone.java
@@ -6,8 +6,8 @@ import java.util.Comparator;
 public class SortedByPhone implements Comparator<Contact> {
     @Override
     public int compare(Contact o1, Contact o2) {
-        int phone1 = o1.getPhone();
-        int phone2 = o2.getPhone();
+        Long phone1 = Long.valueOf(o1.getPhone());
+        Long phone2 = Long.valueOf(o2.getPhone());
 
         if (phone1 > phone2) {
             return 1;
